django-API/requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
blessed==1.15.0
botocore==1.12.193
cached-property==1.5.1
cement==2.8.2
certifi==2019.6.16
chardet==3.0.4
colorama==0.3.9
Django==2.2.3
django-environ==0.4.5
django-extensions==2.2.1
django-filter==2.2.0
django-filters==0.2.1
django-fsm==2.6.1
djangorestframework==3.10.1
docker==3.7.3
docker-compose==1.23.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
factory-boy==2.12.0
Faker==2.0.0
future==0.16.0
gunicorn==19.9.0
idna==2.7
importlib-metadata==0.18
jmespath==0.9.4
jsonschema==2.6.0
marshmallow==2.19.5
more-itertools==7.2.0
packaging==19.0
pathspec==0.5.9
pluggy==0.12.0
psycopg2==2.8.3
py==1.8.0
pyparsing==2.4.1
pytest==5.0.1
pytest-django==3.5.1
python-dateutil==2.8.0
pytz==2019.1
PyYAML==3.13
requests==2.20.1
semantic-version==2.5.0
six==1.12.0
sqlparse==0.3.0
termcolor==1.1.0
text-unidecode==1.2
texttable==0.9.1
urllib3==1.24.3
uWSGI==2.0.18
wcwidth==0.1.7
websocket-client==0.56.0
zipp==0.5.2
